--- domashnee-zadanie-no7-dianamukhametgalina-master/.appveyor.yaml
@@ -1,6 +1,0 @@
-image: Visual Studio 2015
-
-
-
-build:
-  verbosity: minimal
