--- domashnee-zadanie-no7-dianamukhametgalina-master/Snake/MainWindow.xaml.cs
@@ -1,431 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-using System.Windows.Threading;
-
-namespace Snake
-{
-    /// <summary>
-    /// Логика взаимодействия для MainWindow.xaml
-    /// </summary>
-    public partial class MainWindow : Window
-    {
-        //Поле на котором живет змея
-        Entity field;
-        // голова змеи
-        Head head;
-        // вся змея
-        List<PositionedEntity> snake;
-        // яблоко
-        Apple apple;
-        // яблоки для особого режима
-        List<Apple> Apples;
-        //количество очков
-        int score;
-        //таймер по которому 
-        DispatcherTimer moveTimer;
-        //специальный таймер
-        DispatcherTimer specialTimer;
-        //флаг если активен специальный режим
-        bool specialModOn = false;
-
-        //конструктор формы, выполняется при запуске программы
-        public MainWindow()
-        {
-            InitializeComponent();
-
-            snake = new List<PositionedEntity>();
-            //создаем поле 300х300 пикселей
-            field = new Entity(600, 600, "pack://application:,,,/Resources/snake.png");
-
-            //создаем таймер срабатывающий раз в 300 мс
-            moveTimer = new DispatcherTimer();
-            moveTimer.Interval = new TimeSpan(0, 0, 0, 0, 300);
-            moveTimer.Tick += new EventHandler(moveTimer_Tick);
-
-            //создаем таймер срабатывающий раз в 10000 мс
-            specialTimer = new DispatcherTimer();
-            specialTimer.Interval = new TimeSpan(0, 0, 0, 0, 10000);
-            specialTimer.Tick += new EventHandler(specialTimer_Tick);
-        }
-
-        //метод перерисовывающий экран
-        private void UpdateField()
-        {
-            //обновляем положение элементов змеи
-            foreach (var p in snake)
-            {
-                Canvas.SetTop(p.image, p.y);
-                Canvas.SetLeft(p.image, p.x);
-            }
-            //обновляем положение яблока
-            Canvas.SetTop(apple.image, apple.y);
-            Canvas.SetLeft(apple.image, apple.x);
-            if (specialModOn)
-            {
-                foreach (var a in Apples)
-                {
-                    //обновляем положение яблок
-                    Canvas.SetTop(a.image, a.y);
-                    Canvas.SetLeft(a.image, a.x);
-                }
-            }
-
-            //обновляем количество очков
-            lblScore.Content = String.Format("{0}000", score);
-        }
-
-        //обработчик тика таймера. Все движение происходит здесь
-        void moveTimer_Tick(object sender, EventArgs e)
-        {
-            //в обратном порядке двигаем все элементы змеи
-            foreach (var p in Enumerable.Reverse(snake))
-            {
-                p.move();
-            }
-
-            //проверяем, что голова змеи не врезалась в тело
-            foreach (var p in snake.Where(x => x != head))
-            {
-                //если координаты головы и какой либо из частей тела совпадают
-                if (p.x == head.x && p.y == head.y)
-                {
-                    //мы проиграли
-                    moveTimer.Stop();
-                    tbGameOver.Visibility = Visibility.Visible;
-                    return;
-                }
-            }
-
-            //проверяем, что голова змеи не вышла за пределы поля
-            if (head.x < 40 || head.x >= 540 || head.y < 40 || head.y >= 540)
-            {
-                //мы проиграли
-                moveTimer.Stop();
-                tbGameOver.Visibility = Visibility.Visible;
-                return;
-            }
-            if (score == 5 && !specialModOn) // количество яблок которое необходимо съесть чтобы активировать специальный режим
-            {
-                specialModOn = true;
-                specialTimer.Start();
-                apple.remove();
-                int k = 0;
-                for (int x = 40; x < 40 * 14; x += 80)
-                {
-                    for (int y = 40; y < 40 * 14; y += 80)
-                    {
-                        Apples[k].x = x;
-                        Apples[k].y = y;
-                        k++;
-                    }
-                }
-                foreach (var a in Apples)
-                {
-                    a.moveAway();
-                }
-                UpdateField();
-            }
-
-            if (!specialModOn)
-            {
-                //проверяем, что голова змеи врезалась в яблоко
-                if (head.x == apple.x && head.y == apple.y)
-                {
-                    //увеличиваем счет
-                    score++;
-                    //двигаем яблоко на новое место
-                    apple.move();
-                    // добавляем новый сегмент к змее
-                    var part = new BodyPart(snake.Last());
-                    canvas1.Children.Add(part.image);
-                    snake.Add(part);
-                }
-            }
-            else
-            {
-                foreach (var a in Apples)
-                {
-                    if (head.x == a.x && head.y == a.y)
-                    {
-                        //увеличиваем счет
-                        score++;
-                        a.remove();
-                        // добавляем новый сегмент к змее
-                        var part = new BodyPart(snake.Last());
-                        canvas1.Children.Add(part.image);
-                        snake.Add(part);
-                    }
-                }
-            }
-            //перерисовываем экран
-            UpdateField();
-        }
-
-        void specialTimer_Tick(object sender, EventArgs e)
-        {
-            specialModOn = false;
-            specialTimer.Stop();
-            foreach (var a in Apples)
-            {
-                a.remove();
-            }
-            apple.move();
-            foreach (var a in Apples)
-            {
-                //обновляем положение яблок
-                Canvas.SetTop(a.image, a.y);
-                Canvas.SetLeft(a.image, a.x);
-            }
-            UpdateField();
-        }
-
-        // Обработчик нажатия на кнопку клавиатуры
-        private void Window_KeyDown(object sender, KeyEventArgs e)
-        {
-            switch (e.Key)
-            {
-                case Key.Up:
-                    head.direction = Head.Direction.UP;
-                    break;
-                case Key.Down:
-                    head.direction = Head.Direction.DOWN;
-                    break;
-                case Key.Left:
-                    head.direction = Head.Direction.LEFT;
-                    break;
-                case Key.Right:
-                    head.direction = Head.Direction.RIGHT;
-                    break;
-            }
-        }
-
-        // Обработчик нажатия кнопки "Start"
-        private void button1_Click(object sender, RoutedEventArgs e)
-        {
-            // обнуляем счет
-            score = 0;
-            // обнуляем змею
-            snake.Clear();
-            // очищаем канвас
-            canvas1.Children.Clear();
-            // скрываем надпись "Game Over"
-            tbGameOver.Visibility = Visibility.Hidden;
-            // добавляем поле на канвас
-            canvas1.Children.Add(field.image);
-            // создаем новое яблоко и добавлем его
-            apple = new Apple(snake);
-            canvas1.Children.Add(apple.image);
-            // создаём яблоки для особого режима
-            Apples = new List<Apple>(64);
-            for (int i = 0; i < 64; i++)
-            {
-                Apples.Add(new Apple(snake));
-            }
-            foreach (var a in Apples)
-                a.remove();
-            foreach (var a in Apples)
-                canvas1.Children.Add(a.image);
-            foreach (var a in Apples)
-            {
-                //обновляем положение яблок
-                Canvas.SetTop(a.image, a.y);
-                Canvas.SetLeft(a.image, a.x);
-            }
-            // создаем голову
-            head = new Head();
-            snake.Add(head);
-            canvas1.Children.Add(head.image);
-            //запускаем таймер
-            moveTimer.Start();
-            UpdateField();
-
-        }
-
-        public class Entity
-        {
-            protected int m_width;
-            protected int m_height;
-
-            Image m_image;
-            public Entity(int w, int h, string image)
-            {
-                m_width = w;
-                m_height = h;
-                m_image = new Image();
-                m_image.Source = (new ImageSourceConverter()).ConvertFromString(image) as ImageSource;
-                m_image.Width = w;
-                m_image.Height = h;
-
-            }
-
-            public Image image
-            {
-                get
-                {
-                    return m_image;
-                }
-            }
-        }
-
-        public class PositionedEntity : Entity
-        {
-            protected int m_x;
-            protected int m_y;
-            public PositionedEntity(int x, int y, int w, int h, string image)
-                : base(w, h, image)
-            {
-                m_x = x;
-                m_y = y;
-            }
-
-            public virtual void move() { }
-
-            public int x
-            {
-                get
-                {
-                    return m_x;
-                }
-                set
-                {
-                    m_x = value;
-                }
-            }
-
-            public int y
-            {
-                get
-                {
-                    return m_y;
-                }
-                set
-                {
-                    m_y = value;
-                }
-            }
-        }
-
-        public class Apple : PositionedEntity
-        {
-            List<PositionedEntity> m_snake;
-            public Apple(List<PositionedEntity> s)
-                : base(0, 0, 40, 40, "pack://application:,,,/Resources/apple.png")
-            {
-                m_snake = s;
-                move();
-            }
-
-            public override void move()
-            {
-                Random rand = new Random();
-                do
-                {
-                    x = rand.Next(13) * 40 + 40;
-                    y = rand.Next(13) * 40 + 40;
-                    bool overlap = false;
-                    foreach (var p in m_snake)
-                    {
-                        if (p.x == x && p.y == y)
-                        {
-                            overlap = true;
-                            break;
-                        }
-                    }
-                    if (!overlap)
-                        break;
-                } while (true);
-
-            }
-            public void moveAway()
-            {
-                foreach (var p in m_snake)
-                {
-                    if (p.x == x && p.y == y)
-                    {
-                        x = 10000;
-                        y = 10000;
-                        break;
-                    }
-                }
-            }
-            public void remove()
-            {
-                x = 10000;
-                y = 10000;
-            }
-        }
-
-        public class Head : PositionedEntity
-        {
-            public enum Direction
-            {
-                RIGHT, DOWN, LEFT, UP, NONE
-            };
-
-            Direction m_direction;
-
-            public Direction direction
-            {
-                set
-                {
-                    m_direction = value;
-                    RotateTransform rotateTransform = new RotateTransform(90 * (int)value);
-                    image.RenderTransform = rotateTransform;
-                }
-            }
-
-            public Head()
-                : base(280, 280, 40, 40, "pack://application:,,,/Resources/head.png")
-            {
-                image.RenderTransformOrigin = new Point(0.5, 0.5);
-                m_direction = Direction.NONE;
-            }
-
-            public override void move()
-            {
-                switch (m_direction)
-                {
-                    case Direction.DOWN:
-                        y += 40;
-                        break;
-                    case Direction.UP:
-                        y -= 40;
-                        break;
-                    case Direction.LEFT:
-                        x -= 40;
-                        break;
-                    case Direction.RIGHT:
-                        x += 40;
-                        break;
-                }
-            }
-        }
-
-        public class BodyPart : PositionedEntity
-        {
-            PositionedEntity m_next;
-            public BodyPart(PositionedEntity next)
-                : base(next.x, next.y, 40, 40, "pack://application:,,,/Resources/body.png")
-            {
-                m_next = next;
-            }
-
-            public override void move()
-            {
-                x = m_next.x;
-                y = m_next.y;
-            }
-        }
-    }
-}
